refactor: major: wip: moving downsampler to own class, turning DataView into virtual class
Path when aksked to copy over the data to a DataView will call the
"allocate" method. The idea is that the OpenGL code will subclass
DataView and supply an OpenGL allocator method. This way the
Simulation class can call the copy method, taking care of locking
the data structure but not have to know any OpenGL internals