rewrite: Start of major rewrite
The new architecture involves separating out simulation
and visualization completely. In this vision the simulator
will produce a simulation file on disk and the visualizer
will load this file from disk to display it.

This will force us to make a cleaner separation between
simulation and display and greatly simplify the code. In the
event that, in practice, this slows down operation unacceptably
we can then think about ways to pipe the data from the simulator
to the display without actually writing to disk.