Add color and real_body to metadata
This allows us to set a color code for the display
and to skip plotting of barycenters.

The next step will be to figure out how to merge the
metadata for OrreryBody and Spaceship (TBD) such that
we can store it uniformly in the Buffer.

We'll avoid having different types for OrreryBody and
Spaceship because that'll simplify the code a lot.