Force saving of final simulation step to buffer
The fractal downsampler may or may not save the final
step of the simulation. This can cause the display to
look ugly/inaccurate. We now force the buffer to save
the last data point.

Informal profiling (orrery running the complete solar
system for one year with 6min time step) confirms that
this additional check does not add any measurable
compute time.