Remove intermediary vector of spaceship bodies
It is more convenient for integration and flightplan actions
to operate directly on the "body" field of the flightplan
rather than to make an additional copy in Spaceships.

The potential downside to this is a loss of locality due to
the larger size of the FlightPlan structure compared to the
Body structure.