Make Orrery reloading smarter
Orrery now not reloaded if the new range is within the old range.
This considerably speeds up iterations with big orreries
(e.g. the full solarsystem) when all we do is shorten the duration of
the simulation.

(refactor) Relevant computations that do not need to be
implemented as member functions are now moved to
stand-alone functions