Implement function to compute Orrery Body velocities
Sometimes we'll need the velocities of the Orrery bodies too.
The immediate case is for when we initialize a ship in orbit
round a body (cool new feature!) we need to know the body's
speed, since the orbital velocity is relative to that.

In the implemented function we invoke get_orrery_at twice
with the given interval and then use the position difference
to fill out the velocity.