Implement on-demand fractal sampler forcing and buffer flushing
This allows us to force the fractal sampler to save the last data point
on demand, allowing us to save this point to the buffer.

This solves the problem of jerky traces (fairly straight trajectories).
It also elegantly allows us to close out the simulation so we can
get rid of the awkward "force" parameter that we added just for this.