Camera can now be orbited like a chase camera
We now use an azimuth, elevation formulation to determine
where the camera looks at. This formulation makes it easier
for me to think about billboards.

Also added a "center" paramter that will allow us to point
the camera at different objects in the simulation.

Hardcoded into the camera is a correction for the Earth's
axial tilt, to bring the J2000 frame roughly in alignment
with the ecliptic. This should be turned into a variable,
perhaps a general correction that can be specificed or
obtained from the Orrery model.