docs: Add note on new conceptual design
"This then led to the idea that everything parsed
in the scenario files related to the Ships, from static
initial conditions, like the fuel capacity of a ship to
spacecraft actions should be represented in the abstract
as functions to be executed by the `Simulator`. ...

This became very satisfying because the `orbiting`
property of a ship - which was an initial condition but
which required world state had to be awkwardly executed
at the start of simulation run as an exception. This would now
smoothly fit in with the over all philosophy."