Add begin_s and end_s as SpkOrrery properties
This allows us to be more accurate in determining if
a new configuration really requires an Orrery reload.
Previously we would compare two successive configurations.
In the old way if we went from range1 to range2 (smaller)
and then back to range1 we would needlessly reload the
Orrery on the last transition because range1 > range2
even though the loaded Orrery could cover that range.

This also allows us to consistently load the same range
from multiple orreries.