major-refactor: Restructure Scenario PT I
This moves around some of the data structures to
make it conceptually easier to think of spacecraft
actions. It prevents the wierd circular dependencies
we were running into (unknowingly) with the previous
organization.

Now the Configuration handles the loading of
scenario files, while the Scenario carries the
Orrery and the Ships and FlightPlans.

A new structure called State carries the
Orrery and Ships state and can be passed to actions
to allow them to do their thing.

It's also planned that actions will form their
own data structure separate from the spaceships.
This makes for a nice separation of concerns
and we can get rid of the Ship class and actions
can take their place in the Simulator as a Simulator
concern.

This also makes it much easier to think of things
like initially placing a ship in orbit as just
another action that takes place at the first time
step and we no longer have the awkwardness
with such initializers. This also will make the
parser cleaner.

This code is functional but incomplete. It will compile
and run, but only the Orrery will be loaded, no ships
will be created.

We have to work on re-creating the FlightPlanAction
parser and the Actions.