Improve Orrery loading
The original Orrery loading code sorts the order of ephemera after each load.
This leads to issues when an orrery is loaded which contains bodies which
refer to centers not yet loaded. The old code would go into an infinite
loop.

- This is not an issue with the main-line bsp files for the solarsystem
  which always include the sun
- This is an issue with custom .bsp like ast343de430.bsp which refers
  everything to 10 but does not include 10.
- We can not assume all .bsps will be as well behaved as the officail
  ones from NASA

This commit solves this problem by only sorting the Orrery after all
files have been loaded (adds a finalization step). During the sorting
we add a check to the sort that will flag the condition where we
have ended up loading an Orrery that has bodies that refer to centers
that do not exist. We log an error in such a case and move on.

Fixes #11