performance: Remove unnecessary allocation in tight loop
This less than 10 character change causes a 5x speed up in the
code. I was basically copying the Elements repeatedly in a tight
loop. By changing the variable `e` to be a const reference these
allocations/deallocations went away resulting in a scary speedup.

I found this simply by running Instruments/Time Profile once. I
was surprised to find a new in the middle of it all, taking up
so much time and so I found it.