Display can use last (unsampled) point in buffer
Display polling no longer forces the current state to be
sampled in the buffer. Instead the display is able to see
and use the unsampled point at the end of the buffer. The
unsampled point is correctly handled during array updation.

Todo: refactor this code to be nicer and more elegant.

This fixes #8 (Display poll should not force simulation sampling)