Add name to spkid_t
For computation we just want the id, but
for visualization we often want the name too.
It is pleasant to carry the name with the id.

Also, get rid of std::optional in BodyTree.
In usage, we interpret an empty or invalid BodyTree
as the SSB (0) so we just remove this and ensure the
BodyTree always has the SSB as an option.