Implement body markers!
The markers are billboard circles. The default is to be
scaled according to the size of the body. A non-scaling
mode has also been implemented. The next set of commits
will add a set of unscaled markers to the view and
options on the view to show/hide any of the components.