Improve appearance of markers and paths
1. Markers are now yellow points
2. Trajectories are blue lines
3. Scaled models of bodies are as they were
4. Classes have been named more descriptively
5. Much of the code has been split out into .cpp files
5. The colors are set in style.hpp for convenient adjustment

The display looks much more elegant now. The earlier markers,
which were circles with a small non-zero size, looked clunky.
Making them points made busy displays, such as the asteroids
scenario, look really nice. The problem with the points was
that we no longer got the nice markers gliding over the paths
when we advanced time. Making the markers yellow and the paths
blue makes the points stand out and now the animation looks
quite nice. In a busy scene like the asteroids we get a nicely
visible cloud of yellow points over a blue background.