refactor: Separate out Rock and Ship
Instead of having one struct which is a union of the
characteristics we need for Rocks and Ships, we
split them into separate objects.

In the simulator we were already considering
these two objects spearately even though we had a single
array storing the common type. We has to come up with
an awkward way of separating the two classes (by reserving
the first part of the vector for Rocks and the second
part for Ships).

In the display it is much simpler, however, to treat these
as homogenous objects to be displayed. This change should
not make the conversion from simulation to display any more
complex than it already is, with the added benefit we no
longer have the awkward indexing trick, and we can give
the display access to separate Rock and Ship properties
which will come in useful in case we want a more detailed
display (labeling) later.