refactor: Remove redundant base class
We orginally started in this way because we
envisioned different Orrery engines. The thought
was that supporting the SPK models would take
time and it was going to be convenient to start
with toy solar system models to allow work on the
rest of the simulator. The SPK work progressed
ahead of schedule and now we don't really have
any need for switchable orrery engines.