Make lockable data semantics prettier
We can now use structured bindings to extract out the
data and the lock. It looks much nicer this way. It's
also cool to learn that structured bindings "just work"
on classes - they extract public members in order.

We need to do a systematic investigation on why std:pair
fails even though it looks exactly the same to me.