Implement better sampled/unsampled mechanism
The new implementation uses a single vector. We use
a flag to indicate whether the last point is sampled
or an unsampled point. All our operations are greatly
simplified by this modification. We no longer have to
consider the std::optional last point. This has the
added benefit that our interpolation functions now
work properly, considering this last (virtual) point!

We should note here that all this complexity is simply
due to the fact that we like to display the state of
an unfinished simulation - but this feature is so cool
we tolerate the extra complexity it creates across the
codebase.