Implement strong time types + tests
Time conversions are vexing. I chose
formulae from https://www.fourmilab.ch/documents/calendar/
and always convert to and from Gregorian.
For the most part this agrees well with the JPL calculator
(https://ssd.jpl.nasa.gov/tc.cgi#top) for later dates
but starts to drift with earlier dates. For earlier
dates JPL seems to use the Julian Calendar (for
JD 2287184.5 for example) possibly because JPL
tries to output calendar dates using the actual
calendar in use at that time.

The formulae given in Wikipedia are an even poorer match.

I intend to use the simulator for dates in the future
(21st, 22nd century) so this is not an issue.

Internally Groho uses seconds since J2000 as that is
what is used by JPL spice kernels, so this is only an
issue when entering and retrieving dates in the
Gregorian calender, and only an issue before the 16th
century.