Path to kernel now resolved relative to file
Instead of a dumb string, the kernel token is
now furnished with a path relative to the
input file in which it is contained. This
makes kernel loading work properly.

In order to implement this, each Line object
now carries a parent file path object rather
than a simple string.
Implement path awareness in kernel entry