Working on skeleton of charting code
The code uses watchdog and FileLock.
Watching the plotting file for changes
warks marvelously. However, watching
the data directory for changes does
not work so well because there are
numerous writes to the data directory
during a simulation and this results in
a queue of multiple calls to the callback.

We need to think of a better mechanism to
monitor for simulation reruns.

The FileLock code works marvelously with
the C++ lock file code.

Some time was spent figuring out how to
watch a file rather than a whole directory
with watchdog.